+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -244,7 +244,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K7H8V1M0PJS0PEXBBAN9Z51V"),
+  _id: z.string().default("01K7HH3F8Q95PCFR9CPNEFX6BF"),
   display_name: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$inboundSchema,
@@ -288,7 +288,7 @@ export const UpdateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01K7H8V1M0PJS0PEXBBAN9Z51V"),
+  id: z.string().default("01K7HH3F8Q95PCFR9CPNEFX6BF"),
   displayName: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$outboundSchema,