+++ packages/orq-rc/RELEASES.md
@@ -4239,3 +4239,13 @@ Based on:
 - [typescript v3.14.0-rc.24] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.24] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.24 - packages/orq-rc
+
+## 2025-10-14 14:15:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.3 (2.723.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.25] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.25] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.25 - packages/orq-rc